git-svn-id: https://subversion.imagemagick.org/subversion/ImageMagick/trunk@16611 aa41f4f7-0bf4-0310-aa73-e5a19afd5a74

+++ MagickCore/effect.c
@@ -1611,8 +1611,9 @@ MagickExport Image *KuwaharaImage(const Image *image,const double radius,
           }
       }
       status=InterpolatePixelChannels(gaussian_image,image_view[0],
-        kuwahara_image,UndefinedInterpolatePixel,quadrant[l].x+
-        quadrant[l].width/2.0,quadrant[l].y+quadrant[l].height/2.0,q,exception);
+        kuwahara_image,UndefinedInterpolatePixel,(double) quadrant[l].x+
+        quadrant[l].width/2.0,(double) quadrant[l].y+quadrant[l].height/2.0,q,
+        exception);
       q+=GetPixelChannels(kuwahara_image);
     }
     if (SyncCacheViewAuthenticPixels(kuwahara_view,exception) == MagickFalse)
